+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] Pricing
+
+- score matrices
+- starting with historical initially, will then test the rest. Shoot the sheriff
+
 ## [1.4.3]
 
 - readme fix